requirements.txt
fastapi==0.75.0
uvicorn==0.15.0
langchain
langchain-openai
openai==0.11.2
python-dotenv==0.19.2
pytest==6.2.5
httpx==0.22.0